assignment/ai_tool/requirements.txt
streamlit
google-generativeai
sqlalchemy
pandas
python-dotenv
pymysql